pyproject.toml
[project]
name = "nice-prompt"
version = "0.1.0"
description = ""
authors = [
    {name = "Michael Ikemann",email = "michael.ikemann@gmail.com"}
]
readme = "README.md"
requires-python = "^3.12"
dependencies = [
    "nicegui (>=3.3.1,<4.0.0)",
    "yfinance (>=0.2.66,<0.3.0)",
    "pillow (>=12.0.0,<13.0.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.poetry.group.dev.dependencies]
pytest = "^9.0.1"
pytest-asyncio = "^1.3.0"
httpx = "^0.28.1"
tiktoken = "^0.12.0"
pyyaml = "^6.0.3"

